+++ CHANGELOG.md
@@ -4,6 +4,12 @@ We aim to adhere to [semantic versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Version 0.4.1 - Unreleased
 
+### Changed
+* Attaching directories now has mirrors that resolve to exact path matches on the guest and paths relative to root.
+
+### Fixed
+* Attaching directories now uses consistent guest locations between different attach modes 
+
 
 ## [Version 0.4.0] - Released 2026-03-27